+++ chalice/awsclient.py
@@ -129,6 +129,21 @@ class TypedAWSClient(object):
             vpc_config['SecurityGroupIds'] = security_group_ids
         return vpc_config
 
+    def publish_layer(self, layer_name, zip_contents, runtime):
+        # type: (str, bytes, str) -> str
+        try:
+            return self._client('lambda').publish_layer_version(
+                LayerName=layer_name,
+                Content={'ZipFile': zip_contents},
+                CompatibleRuntimes=[runtime])['LayerVersionArn']
+        except _REMOTE_CALL_ERRORS as e:
+            context = LambdaErrorContext(
+                layer_name,
+                'publish_layer_version',
+                len(zip_contents)
+            )
+            raise self._get_lambda_code_deployment_error(e, context)
+
     def create_function(self,
                         function_name,               # type: str
                         role_arn,                    # type: str