+++ chalice/deploy/deployer.py
@@ -81,7 +81,7 @@ is used as the key in the ``deployed.json`` dictionary.
 
 
 """
-
+# pylint: disable=too-many-lines
 import json
 import os
 import textwrap
@@ -382,6 +382,16 @@ class ApplicationGraphBuilder(object):
         # type: (Config, str) -> models.Application
         resources = []  # type: List[models.Model]
         deployment = models.DeploymentPackage(models.Placeholder.BUILD_STAGE)
+        resources.append(
+            models.LambdaLayer(
+                resource_name='layer',
+                layer_name='%s-%s-%s' % (
+                    config.app_name or 'chalice',
+                    config.chalice_stage, 'layer'),
+                runtime=config.lambda_python_version,
+                deployment_package=models.DeploymentPackage(
+                    models.Placeholder.BUILD_STAGE)))
+
         for function in config.chalice_app.pure_lambda_functions:
             resource = self._create_lambda_model(
                 config=config, deployment=deployment,