+++ chalice/deploy/models.py
@@ -150,6 +150,18 @@ class ManagedIAMRole(IAMRole, ManagedModel):
 
 
 @attrs
+class LambdaLayer(ManagedModel):
+    resource_type = 'lambda_layer'
+    layer_name = attrib()             # type: str
+    runtime = attrib()                # type: str
+    deployment_package = attrib()     # type: DeploymentPackage
+
+    def dependencies(self):
+        # type: () -> List[Model]
+        return [self.deployment_package]
+
+
+@attrs
 class LambdaFunction(ManagedModel):
     resource_type = 'lambda_function'
     function_name = attrib()          # type: str