+++ tests/functional/cli/test_cli.py
@@ -149,7 +149,8 @@ def test_can_package_with_single_file(runner):
         assert result.exit_code == 0, result.output
         assert os.path.isfile('package.zip')
         with zipfile.ZipFile('package.zip', 'r') as f:
-            assert sorted(f.namelist()) == ['deployment.zip', 'sam.json']
+            assert sorted(f.namelist()) == [
+                'deployment.zip', 'layer-deployment.zip', 'sam.json']
 
 
 def test_debug_flag_enables_logging(runner):